Convert Dsp to a common DLL library
