+++ Builds/VisualStudio2013/Plugins/Plugins.sln
@@ -11,6 +11,9 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "KWIKFormat", "KWIKFormat\KW
 	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "FilterNode", "FilterNode\FilterNode.vcxproj", "{7B23828E-559F-4AD2-B75D-D05786F6329C}"
+	ProjectSection(ProjectDependencies) = postProject
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B} = {1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}
+	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NetworkEvents", "NetworkEvents\NetworkEvents.vcxproj", "{D5D58DAC-582B-4F39-9385-E1814D56BCA0}"
 EndProject
@@ -54,6 +57,8 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "OpenEphysHDF5Lib", "CommonLibs\OpenEphysHDF5Lib\OpenEphysHDF5Lib.vcxproj", "{F250DB70-6E3E-408C-BD9E-1483D574499D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AA470E62-74F7-4AD0-BAF9-D88B68882D43}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DspLib", "CommonLibs\DspLib\DspLib.vcxproj", "{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -305,11 +310,24 @@ Global
 		{F250DB70-6E3E-408C-BD9E-1483D574499D}.Release|Win32.Build.0 = Release|Win32
 		{F250DB70-6E3E-408C-BD9E-1483D574499D}.Release|x64.ActiveCfg = Release|x64
 		{F250DB70-6E3E-408C-BD9E-1483D574499D}.Release|x64.Build.0 = Release|x64
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Debug|Mixed Platforms.Build.0 = Debug|Win32
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Debug|Win32.ActiveCfg = Debug|Win32
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Debug|Win32.Build.0 = Debug|Win32
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Debug|x64.ActiveCfg = Debug|x64
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Debug|x64.Build.0 = Debug|x64
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Release|Mixed Platforms.ActiveCfg = Release|Win32
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Release|Mixed Platforms.Build.0 = Release|Win32
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Release|Win32.ActiveCfg = Release|Win32
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Release|Win32.Build.0 = Release|Win32
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Release|x64.ActiveCfg = Release|x64
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{F250DB70-6E3E-408C-BD9E-1483D574499D} = {B49D0C12-70A8-4762-A4FE-74FA3D16FFDD}
+		{1DD9DB7A-A9D8-4982-9FA3-1AEAFEC5FC1B} = {B49D0C12-70A8-4762-A4FE-74FA3D16FFDD}
 	EndGlobalSection
 EndGlobal
