+++ Source/Plugins/CommonLibs/DspLib/ChebyshevI.cpp
