+++ Source/Plugins/CommonLibs/DspLib/ChebyshevI.h
@@ -33,84 +33,55 @@ THE SOFTWARE.
 
 *******************************************************************************/
 
-#ifndef DSPFILTERS_BESSEL_H
-#define DSPFILTERS_BESSEL_H
+#ifndef DSPFILTERS_CHEBYSHEVI_H
+#define DSPFILTERS_CHEBYSHEVI_H
 
 #include "Common.h"
 #include "Cascade.h"
 #include "Design.h"
 #include "Filter.h"
 #include "PoleFilter.h"
-#include "RootFinder.h"
 
 namespace Dsp
 {
 
 /*
- * Filters with Bessel response characteristics
+ * Filters with Chebyshev response characteristics
  *
  */
 
-namespace Bessel
+namespace ChebyshevI
 {
-
-// A Workspace is necessary to find roots
-
-struct WorkspaceBase
-{
-    WorkspaceBase(RootFinderBase* rootsBase)
-        : roots(*rootsBase)
-    {
-    }
-
-    RootFinderBase& roots;
-
-private:
-    WorkspaceBase(WorkspaceBase&);
-    WorkspaceBase& operator= (WorkspaceBase&);
-};
-
-template <int MaxOrder>
-struct Workspace : WorkspaceBase
-{
-    Workspace()
-        : WorkspaceBase(&m_roots)
-    {
-    }
-
-private:
-    RootFinder <MaxOrder> m_roots;
-};
-
-//------------------------------------------------------------------------------
 
 // Half-band analog prototypes (s-plane)
 
-class AnalogLowPass : public LayoutBase
+class COMMON_LIB AnalogLowPass : public LayoutBase
 {
 public:
     AnalogLowPass();
 
     void design(const int numPoles,
-                WorkspaceBase* w);
+                double rippleDb);
 
 private:
     int m_numPoles;
+    double m_rippleDb;
 };
 
 //------------------------------------------------------------------------------
 
-class AnalogLowShelf : public LayoutBase
+class COMMON_LIB AnalogLowShelf : public LayoutBase
 {
 public:
     AnalogLowShelf();
 
     void design(int numPoles,
                 double gainDb,
-                WorkspaceBase* w);
+                double rippleDb);
 
 private:
     int m_numPoles;
+    double m_rippleDb;
     double m_gainDb;
 };
 
@@ -118,47 +89,66 @@ private:
 
 // Factored implementations to reduce template instantiations
 
-struct LowPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB LowPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double cutoffFrequency,
-               WorkspaceBase* w);
+               double rippleDb);
 };
 
-struct HighPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB HighPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double cutoffFrequency,
-               WorkspaceBase* w);
+               double rippleDb);
 };
 
-struct BandPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB BandPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double centerFrequency,
                double widthFrequency,
-               WorkspaceBase* w);
+               double rippleDb);
 };
 
-struct BandStopBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB BandStopBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double centerFrequency,
                double widthFrequency,
-               WorkspaceBase* w);
+               double rippleDb);
 };
 
-struct LowShelfBase : PoleFilterBase <AnalogLowShelf>
+struct COMMON_LIB LowShelfBase : PoleFilterBase <AnalogLowShelf>
 {
     void setup(int order,
                double sampleRate,
                double cutoffFrequency,
                double gainDb,
-               WorkspaceBase* w);
+               double rippleDb);
+};
+
+struct COMMON_LIB HighShelfBase : PoleFilterBase <AnalogLowShelf>
+{
+    void setup(int order,
+               double sampleRate,
+               double cutoffFrequency,
+               double gainDb,
+               double rippleDb);
+};
+
+struct COMMON_LIB BandShelfBase : PoleFilterBase <AnalogLowShelf>
+{
+    void setup(int order,
+               double sampleRate,
+               double centerFrequency,
+               double widthFrequency,
+               double gainDb,
+               double rippleDb);
 };
 
 //------------------------------------------------------------------------------
@@ -170,82 +160,36 @@ struct LowShelfBase : PoleFilterBase <AnalogLowShelf>
 template <int MaxOrder>
 struct LowPass : PoleFilter <LowPassBase, MaxOrder>
 {
-    void setup(int order,
-               double sampleRate,
-               double cutoffFrequency)
-    {
-        Workspace <MaxOrder> w;
-        LowPassBase::setup(order,
-                           sampleRate,
-                           cutoffFrequency,
-                           &w);
-    }
 };
 
 template <int MaxOrder>
 struct HighPass : PoleFilter <HighPassBase, MaxOrder>
 {
-    void setup(int order,
-               double sampleRate,
-               double cutoffFrequency)
-    {
-        Workspace <MaxOrder> w;
-        HighPassBase::setup(order,
-                            sampleRate,
-                            cutoffFrequency,
-                            &w);
-    }
 };
 
 template <int MaxOrder>
-struct BandPass : PoleFilter <BandPassBase, MaxOrder, MaxOrder*2>
+struct BandPass : PoleFilter <BandPassBase, MaxOrder, MaxOrder * 2>
 {
-    void setup(int order,
-               double sampleRate,
-               double centerFrequency,
-               double widthFrequency)
-    {
-        Workspace <MaxOrder> w;
-        BandPassBase::setup(order,
-                            sampleRate,
-                            centerFrequency,
-                            widthFrequency,
-                            &w);
-    }
 };
 
 template <int MaxOrder>
-struct BandStop : PoleFilter <BandStopBase, MaxOrder, MaxOrder*2>
+struct BandStop : PoleFilter <BandStopBase, MaxOrder, MaxOrder * 2>
 {
-    void setup(int order,
-               double sampleRate,
-               double centerFrequency,
-               double widthFrequency)
-    {
-        Workspace <MaxOrder> w;
-        BandStopBase::setup(order,
-                            sampleRate,
-                            centerFrequency,
-                            widthFrequency,
-                            &w);
-    }
 };
 
 template <int MaxOrder>
-struct LowShelf : PoleFilter <LowShelfBase, MaxOrder, MaxOrder*2>
+struct LowShelf : PoleFilter <LowShelfBase, MaxOrder>
 {
-    void setup(int order,
-               double sampleRate,
-               double cutoffFrequency,
-               double gainDb)
-    {
-        Workspace <MaxOrder> w;
-        LowShelfBase::setup(order,
-                            sampleRate,
-                            cutoffFrequency,
-                            gainDb,
-                            &w);
-    }
+};
+
+template <int MaxOrder>
+struct HighShelf : PoleFilter <HighShelfBase, MaxOrder>
+{
+};
+
+template <int MaxOrder>
+struct BandShelf : PoleFilter <BandShelfBase, MaxOrder, MaxOrder * 2>
+{
 };
 
 //------------------------------------------------------------------------------
@@ -261,17 +205,22 @@ struct TypeIBase : DesignBase
 {
     enum
     {
-        NumParams = 3
+        NumParams = 4
     };
 
     static int getNumParams()
     {
-        return 3;
+        return 4;
     }
 
     static const ParamInfo getParamInfo_2()
     {
         return ParamInfo::defaultCutoffFrequencyParam();
+    }
+
+    static const ParamInfo getParamInfo_3()
+    {
+        return ParamInfo::defaultRippleDbParam();
     }
 };
 
@@ -280,78 +229,11 @@ struct TypeI : TypeIBase, FilterClass
 {
     void setParams(const Params& params)
     {
-        FilterClass::setup(int(params[1]), params[0], params[2]);
-    }
-};
-
-struct TypeIIBase : DesignBase
-{
-    enum
-    {
-        NumParams = 4
-    };
-
-    static int getNumParams()
-    {
-        return 4;
-    }
-
-    static const ParamInfo getParamInfo_2()
-    {
-        return ParamInfo::defaultCenterFrequencyParam();
-    }
-
-    static const ParamInfo getParamInfo_3()
-    {
-        return ParamInfo::defaultBandwidthHzParam();
-    }
-};
-
-template <class FilterClass>
-struct TypeII : TypeIIBase, FilterClass
-{
-    void setParams(const Params& params)
-    {
         FilterClass::setup(int(params[1]), params[0], params[2], params[3]);
     }
 };
 
-struct TypeIIIBase : DesignBase
-{
-    enum
-    {
-        NumParams = 4
-    };
-
-    static int getNumParams()
-    {
-        return 4;
-    }
-
-    static const ParamInfo getParamInfo_2()
-    {
-        return ParamInfo::defaultCutoffFrequencyParam();
-    }
-
-    static const ParamInfo getParamInfo_3()
-    {
-        return ParamInfo::defaultGainParam();
-    }
-};
-
-template <class FilterClass>
-struct TypeIII : TypeIIIBase, FilterClass
-{
-    void setParams(const Params& params)
-    {
-        FilterClass::setup(int(params[1]),
-                           params[0],
-                           params[2],
-                           params[3]);
-    }
-};
-
-struct TypeIVBase : DesignBase
+struct TypeIIBase : DesignBase
 {
     enum
     {
@@ -375,7 +257,86 @@ struct TypeIVBase : DesignBase
 
     static const ParamInfo getParamInfo_4()
     {
+        return ParamInfo::defaultRippleDbParam();
+    }
+};
+
+template <class FilterClass>
+struct TypeII : TypeIIBase, FilterClass
+{
+    void setParams(const Params& params)
+    {
+        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4]);
+    }
+};
+
+struct TypeIIIBase : DesignBase
+{
+    enum
+    {
+        NumParams = 5
+    };
+
+    static int getNumParams()
+    {
+        return 5;
+    }
+
+    static const ParamInfo getParamInfo_2()
+    {
+        return ParamInfo::defaultCutoffFrequencyParam();
+    }
+
+    static const ParamInfo getParamInfo_3()
+    {
         return ParamInfo::defaultGainParam();
+    }
+
+    static const ParamInfo getParamInfo_4()
+    {
+        return ParamInfo::defaultRippleDbParam();
+    }
+};
+
+template <class FilterClass>
+struct TypeIII : TypeIIIBase, FilterClass
+{
+    void setParams(const Params& params)
+    {
+        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4]);
+    }
+};
+
+struct TypeIVBase : DesignBase
+{
+    enum
+    {
+        NumParams = 6
+    };
+
+    static int getNumParams()
+    {
+        return 6;
+    }
+
+    static const ParamInfo getParamInfo_2()
+    {
+        return ParamInfo::defaultCenterFrequencyParam();
+    }
+
+    static const ParamInfo getParamInfo_3()
+    {
+        return ParamInfo::defaultBandwidthHzParam();
+    }
+
+    static const ParamInfo getParamInfo_4()
+    {
+        return ParamInfo::defaultGainParam();
+    }
+
+    static const ParamInfo getParamInfo_5()
+    {
+        return ParamInfo::defaultRippleDbParam();
     }
 };
 
@@ -384,7 +345,7 @@ struct TypeIV : TypeIVBase, FilterClass
 {
     void setParams(const Params& params)
     {
-        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4]);
+        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4], params[5]);
     }
 };
 
@@ -398,7 +359,7 @@ struct LowPassDescription
     }
     static const char* getName()
     {
-        return "Bessel Low Pass";
+        return "Chebyshev I Low Pass";
     }
 };
 
@@ -410,7 +371,7 @@ struct HighPassDescription
     }
     static const char* getName()
     {
-        return "Bessel High Pass";
+        return "Chebyshev I High Pass";
     }
 };
 
@@ -422,7 +383,7 @@ struct BandPassDescription
     }
     static const char* getName()
     {
-        return "Bessel Band Pass";
+        return "Chebyshev I Band Pass";
     }
 };
 
@@ -434,7 +395,7 @@ struct BandStopDescription
     }
     static const char* getName()
     {
-        return "Bessel Band Stop";
+        return "Chebyshev I Band Stop";
     }
 };
 
@@ -446,7 +407,31 @@ struct LowShelfDescription
     }
     static const char* getName()
     {
-        return "Bessel Low Shelf";
+        return "Chebyshev I Low Shelf";
+    }
+};
+
+struct HighShelfDescription
+{
+    static Kind getKind()
+    {
+        return kindHighShelf;
+    }
+    static const char* getName()
+    {
+        return "Chebyshev I High Shelf";
+    }
+};
+
+struct BandShelfDescription
+{
+    static Kind getKind()
+    {
+        return kindBandShelf;
+    }
+    static const char* getName()
+    {
+        return "Chebyshev I Band Shelf";
     }
 };
 
@@ -470,40 +455,48 @@ struct OrderBase : TypeClass <FilterClass <MaxOrder> >
 //------------------------------------------------------------------------------
 
 //
-// Gui-friendly Design layer
+// Design filters
 //
 
 template <int MaxOrder>
-struct LowPass : OrderBase <MaxOrder, TypeI, Bessel::LowPass>,
+struct LowPass : OrderBase <MaxOrder, TypeI, ChebyshevI::LowPass>,
         LowPassDescription
 {
 };
 
 template <int MaxOrder>
-struct HighPass : OrderBase <MaxOrder, TypeI, Bessel::HighPass>,
+struct HighPass : OrderBase <MaxOrder, TypeI, ChebyshevI::HighPass>,
         HighPassDescription
 {
 };
 
 template <int MaxOrder>
-struct BandPass : OrderBase <MaxOrder, TypeII, Bessel::BandPass>,
+struct BandPass : OrderBase <MaxOrder, TypeII, ChebyshevI::BandPass>,
         BandPassDescription
 {
 };
 
 template <int MaxOrder>
-struct BandStop : OrderBase <MaxOrder, TypeII, Bessel::BandStop>,
+struct BandStop : OrderBase <MaxOrder, TypeII, ChebyshevI::BandStop>,
         BandStopDescription
 {
 };
 
-/*
- * NOT IMPLEMENTED
- *
- */
 template <int MaxOrder>
-struct LowShelf : OrderBase <MaxOrder, TypeIII, Bessel::LowShelf>,
+struct LowShelf : OrderBase <MaxOrder, TypeIII, ChebyshevI::LowShelf>,
         LowShelfDescription
+{
+};
+
+template <int MaxOrder>
+struct HighShelf : OrderBase <MaxOrder, TypeIII, ChebyshevI::HighShelf>,
+        HighShelfDescription
+{
+};
+
+template <int MaxOrder>
+struct BandShelf : OrderBase <MaxOrder, TypeIV, ChebyshevI::BandShelf>,
+        BandShelfDescription
 {
 };
 
@@ -515,4 +508,3 @@ struct LowShelf : OrderBase <MaxOrder, TypeIII, Bessel::LowShelf>,
 
 #endif
 
-/* This is a test of svn:external */
