+++ Source/Plugins/CommonLibs/DspLib/ChebyshevII.cpp
