+++ Source/Plugins/CommonLibs/DspLib/Custom.h
@@ -57,14 +57,14 @@ namespace Custom
 // Raw filters
 //
 
-struct OnePole : Biquad
+struct COMMON_LIB OnePole : Biquad
 {
     void setup(double scale,
                double pole,
                double zero);
 };
 
-struct TwoPole : Biquad
+struct COMMON_LIB TwoPole : Biquad
 {
     void setup(double scale,
                double poleRho,
