+++ Source/Plugins/CommonLibs/DspLib/Design.cpp
