+++ Source/Plugins/CommonLibs/DspLib/Design.h
