+++ Source/Plugins/CommonLibs/DspLib/Documentation.cpp
