+++ Source/Plugins/CommonLibs/DspLib/Elliptic.cpp
