+++ Source/Plugins/CommonLibs/DspLib/Elliptic.h
@@ -33,8 +33,8 @@ THE SOFTWARE.
 
 *******************************************************************************/
 
-#ifndef DSPFILTERS_CHEBYSHEVII_H
-#define DSPFILTERS_CHEBYSHEVII_H
+#ifndef DSPFILTERS_ELLIPTIC_H
+#define DSPFILTERS_ELLIPTIC_H
 
 #include "Common.h"
 #include "Cascade.h"
@@ -46,109 +46,124 @@ namespace Dsp
 {
 
 /*
- * Filters with Inverse Chebyshev response characteristics
+ * Filters with Elliptic response characteristics
  *
  */
 
-namespace ChebyshevII
+namespace Elliptic
 {
 
-// Half-band analog prototypes (s-plane)
+// Solves for Jacobi elliptics
+class COMMON_LIB Solver
+{
+public:
+    static double ellipticK(double k);
+};
 
-class AnalogLowPass : public LayoutBase
+// Half-band analog prototype (s-plane)
+
+class COMMON_LIB AnalogLowPass : public LayoutBase
 {
 public:
     AnalogLowPass();
 
     void design(const int numPoles,
-                double stopBandDb);
+                double rippleDb,
+                double rolloff);
 
 private:
+    void prodpoly(int sn);
+    void calcfz2(int i);
+    void calcfz();
+    void calcqz();
+    double findfact(int t);
+    double calcsn(double u);
+
+#if 0
+    template<int n>
+    struct CalcArray
+    {
+        double& operator[](size_t index)
+        {
+            assert(index<n);
+            return m_a[index];
+        }
+    private:
+        double m_a[n];
+    };
+#else
+#endif
+
+    double m_p0;
+    double m_q;
+    double m_K;
+    double m_Kprime;
+    double m_e;
+    int m_nin;
+    int m_m;
+    int m_n2;
+    int m_em;
+    double m_zeros[100];
+    double m_c1[100];
+    double m_b1[100];
+    double m_a1[100];
+    double m_d1[100];
+    double m_q1[100];
+    double m_z1[100];
+    double m_f1[100];
+    double m_s1[100];
+    double m_p [100];
+    double m_zw1[100];
+    double m_zf1[100];
+    double m_zq1[100];
+    double m_rootR[100];
+    double m_rootI[100];
+
     int m_numPoles;
-    double m_stopBandDb;
-};
-
-//------------------------------------------------------------------------------
-
-class AnalogLowShelf : public LayoutBase
-{
-public:
-    AnalogLowShelf();
-
-    void design(int numPoles,
-                double gainDb,
-                double stopBandDb);
-
-private:
-    int m_numPoles;
-    double m_stopBandDb;
-    double m_gainDb;
+    double m_rippleDb;
+    double m_rolloff;
 };
 
 //------------------------------------------------------------------------------
 
 // Factored implementations to reduce template instantiations
 
-struct LowPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB LowPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double cutoffFrequency,
-               double stopBandDb);
+               double rippleDb,
+               double rolloff);
 };
 
-struct HighPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB HighPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double cutoffFrequency,
-               double stopBandDb);
+               double rippleDb,
+               double rolloff);
 };
 
-struct BandPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB BandPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double centerFrequency,
                double widthFrequency,
-               double stopBandDb);
+               double rippleDb,
+               double rolloff);
 };
 
-struct BandStopBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB BandStopBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double centerFrequency,
                double widthFrequency,
-               double stopBandDb);
-};
-
-struct LowShelfBase : PoleFilterBase <AnalogLowShelf>
-{
-    void setup(int order,
-               double sampleRate,
-               double cutoffFrequency,
-               double gainDb,
-               double stopBandDb);
-};
-
-struct HighShelfBase : PoleFilterBase <AnalogLowShelf>
-{
-    void setup(int order,
-               double sampleRate,
-               double cutoffFrequency,
-               double gainDb,
-               double stopBandDb);
-};
-
-struct BandShelfBase : PoleFilterBase <AnalogLowShelf>
-{
-    void setup(int order,
-               double sampleRate,
-               double centerFrequency,
-               double widthFrequency,
-               double gainDb,
-               double stopBandDb);
+               double rippleDb,
+               double rolloff);
 };
 
 //------------------------------------------------------------------------------
@@ -177,21 +192,6 @@ struct BandStop : PoleFilter <BandStopBase, MaxOrder, MaxOrder*2>
 {
 };
 
-template <int MaxOrder>
-struct LowShelf : PoleFilter <LowShelfBase, MaxOrder>
-{
-};
-
-template <int MaxOrder>
-struct HighShelf : PoleFilter <HighShelfBase, MaxOrder>
-{
-};
-
-template <int MaxOrder>
-struct BandShelf : PoleFilter <BandShelfBase, MaxOrder, MaxOrder*2>
-{
-};
-
 //------------------------------------------------------------------------------
 
 //
@@ -205,12 +205,12 @@ struct TypeIBase : DesignBase
 {
     enum
     {
-        NumParams = 4
+        NumParams = 5
     };
 
     static int getNumParams()
     {
-        return 4;
+        return 5;
     }
 
     static const ParamInfo getParamInfo_2()
@@ -220,7 +220,12 @@ struct TypeIBase : DesignBase
 
     static const ParamInfo getParamInfo_3()
     {
-        return ParamInfo::defaultStopDbParam();
+        return ParamInfo::defaultRippleDbParam();
+    }
+
+    static const ParamInfo getParamInfo_4()
+    {
+        return ParamInfo::defaultRolloffParam();
     }
 };
 
@@ -229,85 +234,11 @@ struct TypeI : TypeIBase, FilterClass
 {
     void setParams(const Params& params)
     {
-        FilterClass::setup(int(params[1]), params[0], params[2], params[3]);
+        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4]);
     }
 };
 
 struct TypeIIBase : DesignBase
-{
-    enum
-    {
-        NumParams = 5
-    };
-
-    static int getNumParams()
-    {
-        return 5;
-    }
-
-    static const ParamInfo getParamInfo_2()
-    {
-        return ParamInfo::defaultCenterFrequencyParam();
-    }
-
-    static const ParamInfo getParamInfo_3()
-    {
-        return ParamInfo::defaultBandwidthHzParam();
-    }
-
-    static const ParamInfo getParamInfo_4()
-    {
-        return ParamInfo::defaultStopDbParam();
-    }
-};
-
-template <class FilterClass>
-struct TypeII : TypeIIBase, FilterClass
-{
-    void setParams(const Params& params)
-    {
-        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4]);
-    }
-};
-
-struct TypeIIIBase : DesignBase
-{
-    enum
-    {
-        NumParams = 5
-    };
-
-    static int getNumParams()
-    {
-        return 5;
-    }
-
-    static const ParamInfo getParamInfo_2()
-    {
-        return ParamInfo::defaultCutoffFrequencyParam();
-    }
-
-    static const ParamInfo getParamInfo_3()
-    {
-        return ParamInfo::defaultGainParam();
-    }
-
-    static const ParamInfo getParamInfo_4()
-    {
-        return ParamInfo::defaultStopDbParam();
-    }
-};
-
-template <class FilterClass>
-struct TypeIII : TypeIIIBase, FilterClass
-{
-    void setParams(const Params& params)
-    {
-        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4]);
-    }
-};
-
-struct TypeIVBase : DesignBase
 {
     enum
     {
@@ -331,17 +262,17 @@ struct TypeIVBase : DesignBase
 
     static const ParamInfo getParamInfo_4()
     {
-        return ParamInfo::defaultGainParam();
+        return ParamInfo::defaultRippleDbParam();
     }
 
     static const ParamInfo getParamInfo_5()
     {
-        return ParamInfo::defaultStopDbParam();
+        return ParamInfo::defaultRolloffParam();
     }
 };
 
 template <class FilterClass>
-struct TypeIV : TypeIVBase, FilterClass
+struct TypeII : TypeIIBase, FilterClass
 {
     void setParams(const Params& params)
     {
@@ -359,7 +290,7 @@ struct LowPassDescription
     }
     static const char* getName()
     {
-        return "Chebyshev II Low Pass";
+        return "Elliptic Low Pass";
     }
 };
 
@@ -371,7 +302,7 @@ struct HighPassDescription
     }
     static const char* getName()
     {
-        return "Chebyshev II High Pass";
+        return "Elliptic High Pass";
     }
 };
 
@@ -383,7 +314,7 @@ struct BandPassDescription
     }
     static const char* getName()
     {
-        return "Chebyshev II Band Pass";
+        return "Elliptic Band Pass";
     }
 };
 
@@ -395,43 +326,7 @@ struct BandStopDescription
     }
     static const char* getName()
     {
-        return "Chebyshev II Band Stop";
-    }
-};
-
-struct LowShelfDescription
-{
-    static Kind getKind()
-    {
-        return kindLowShelf;
-    }
-    static const char* getName()
-    {
-        return "Chebyshev II Low Shelf";
-    }
-};
-
-struct HighShelfDescription
-{
-    static Kind getKind()
-    {
-        return kindHighShelf;
-    }
-    static const char* getName()
-    {
-        return "Chebyshev II High Shelf";
-    }
-};
-
-struct BandShelfDescription
-{
-    static Kind getKind()
-    {
-        return kindBandShelf;
-    }
-    static const char* getName()
-    {
-        return "Chebyshev II Band Shelf";
+        return "Elliptic Band Stop";
     }
 };
 
@@ -451,55 +346,35 @@ struct OrderBase : TypeClass <FilterClass <MaxOrder> >
 
     }
 };
-
 //------------------------------------------------------------------------------
 
 //
-// Design Filters
+// Design filters
 //
 
 template <int MaxOrder>
-struct LowPass : OrderBase <MaxOrder, TypeI, ChebyshevII::LowPass>,
+struct LowPass : OrderBase <MaxOrder, TypeI, Elliptic::LowPass>,
         LowPassDescription
 {
 };
 
 template <int MaxOrder>
-struct HighPass : OrderBase <MaxOrder, TypeI, ChebyshevII::HighPass>,
+struct HighPass : OrderBase <MaxOrder, TypeI, Elliptic::HighPass>,
         HighPassDescription
 {
 };
 
 template <int MaxOrder>
-struct BandPass : OrderBase <MaxOrder, TypeII, ChebyshevII::BandPass>,
+struct BandPass : OrderBase <MaxOrder, TypeII, Elliptic::BandPass>,
         BandPassDescription
 {
 };
 
 template <int MaxOrder>
-struct BandStop : OrderBase <MaxOrder, TypeII, ChebyshevII::BandStop>,
+struct BandStop : OrderBase <MaxOrder, TypeII, Elliptic::BandStop>,
         BandStopDescription
 {
 };
-
-template <int MaxOrder>
-struct LowShelf : OrderBase <MaxOrder, TypeIII, ChebyshevII::LowShelf>,
-        LowShelfDescription
-{
-};
-
-template <int MaxOrder>
-struct HighShelf : OrderBase <MaxOrder, TypeIII, ChebyshevII::HighShelf>,
-        HighShelfDescription
-{
-};
-
-template <int MaxOrder>
-struct BandShelf : OrderBase <MaxOrder, TypeIV, ChebyshevII::BandShelf>,
-        BandShelfDescription
-{
-};
-
 
 }
