+++ Source/Plugins/CommonLibs/DspLib/Layout.h
