+++ Source/Plugins/CommonLibs/DspLib/Legendre.cpp
