+++ Source/Plugins/CommonLibs/DspLib/Legendre.h
@@ -33,137 +33,170 @@ THE SOFTWARE.
 
 *******************************************************************************/
 
-#ifndef DSPFILTERS_ELLIPTIC_H
-#define DSPFILTERS_ELLIPTIC_H
+#ifndef DSPFILTERS_LEGENDRE_H
+#define DSPFILTERS_LEGENDRE_H
 
 #include "Common.h"
 #include "Cascade.h"
 #include "Design.h"
 #include "Filter.h"
 #include "PoleFilter.h"
+#include "RootFinder.h"
 
 namespace Dsp
 {
 
 /*
- * Filters with Elliptic response characteristics
+ * Filters with Legendre / "Optimum-L" response characteristics
  *
  */
 
-namespace Elliptic
+namespace Legendre
 {
 
-// Solves for Jacobi elliptics
-class Solver
+// Numerical computation of Legendre "Optimum-L" polynomials
+
+class COMMON_LIB PolynomialFinderBase
 {
 public:
-    static double ellipticK(double k);
+    void solve(int n);
+
+    double* coef()
+    {
+        return m_w;
+    }
+
+private:
+    void legendre(double* p, int n);
+
+protected:
+    int m_maxN;
+    double* m_w;
+    double* m_a;
+    double* m_p;
+    double* m_s;
+    double* m_v;
+    double* m_aa;
+    double* m_bb;
 };
 
-// Half-band analog prototype (s-plane)
+template <int maxN>
+class PolynomialFinder : public PolynomialFinderBase
+{
+public:
+    PolynomialFinder()
+    {
+        m_maxN = maxN;
+        m_w = m_ws;
+        m_a = m_as;
+        m_p = m_ps;
+        m_s = m_ss;
+        m_v = m_vs;
+        m_aa = m_aas;
+        m_bb = m_bbs;
+    }
 
-class AnalogLowPass : public LayoutBase
+    void solve(int n)
+    {
+        assert(n <= maxN);
+        PolynomialFinderBase::solve(n);
+    }
+
+private:
+    double m_ws  [2 * maxN + 1];
+    double m_as  [    maxN + 1];
+    double m_ps  [2 * maxN + 1];
+    double m_ss  [2 * maxN + 1];
+    double m_vs  [2 * maxN + 4];
+    double m_aas [    maxN + 1];
+    double m_bbs [    maxN + 1];
+};
+
+//------------------------------------------------------------------------------
+
+// A Workspace is necessary to construct the polynomial and find its roots
+
+struct COMMON_LIB WorkspaceBase
+{
+    WorkspaceBase(PolynomialFinderBase* polyBase,
+                  RootFinderBase* rootsBase)
+        : poly(*polyBase)
+        , roots(*rootsBase)
+    {
+    }
+
+    PolynomialFinderBase& poly;
+    RootFinderBase& roots;
+
+private:
+    WorkspaceBase(WorkspaceBase&);
+    WorkspaceBase& operator= (WorkspaceBase&);
+};
+
+template <int MaxOrder>
+struct Workspace : WorkspaceBase
+{
+    Workspace()
+        : WorkspaceBase(&m_poly, &m_roots)
+    {
+    }
+
+private:
+    PolynomialFinder <MaxOrder> m_poly;
+    RootFinder <MaxOrder * 2> m_roots;
+};
+
+//------------------------------------------------------------------------------
+
+// Half-band analog prototypes (s-plane)
+
+class COMMON_LIB AnalogLowPass : public LayoutBase
 {
 public:
     AnalogLowPass();
 
-    void design(const int numPoles,
-                double rippleDb,
-                double rolloff);
+    void design(const int numPoles, WorkspaceBase* w);
 
 private:
-    void prodpoly(int sn);
-    void calcfz2(int i);
-    void calcfz();
-    void calcqz();
-    double findfact(int t);
-    double calcsn(double u);
-
-#if 0
-    template<int n>
-    struct CalcArray
-    {
-        double& operator[](size_t index)
-        {
-            assert(index<n);
-            return m_a[index];
-        }
-    private:
-        double m_a[n];
-    };
-#else
-#endif
-
-    double m_p0;
-    double m_q;
-    double m_K;
-    double m_Kprime;
-    double m_e;
-    int m_nin;
-    int m_m;
-    int m_n2;
-    int m_em;
-    double m_zeros[100];
-    double m_c1[100];
-    double m_b1[100];
-    double m_a1[100];
-    double m_d1[100];
-    double m_q1[100];
-    double m_z1[100];
-    double m_f1[100];
-    double m_s1[100];
-    double m_p [100];
-    double m_zw1[100];
-    double m_zf1[100];
-    double m_zq1[100];
-    double m_rootR[100];
-    double m_rootI[100];
-
     int m_numPoles;
-    double m_rippleDb;
-    double m_rolloff;
 };
 
 //------------------------------------------------------------------------------
 
 // Factored implementations to reduce template instantiations
 
-struct LowPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB LowPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double cutoffFrequency,
-               double rippleDb,
-               double rolloff);
+               WorkspaceBase* w);
 };
 
-struct HighPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB HighPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double cutoffFrequency,
-               double rippleDb,
-               double rolloff);
+               WorkspaceBase* w);
 };
 
-struct BandPassBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB BandPassBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double centerFrequency,
                double widthFrequency,
-               double rippleDb,
-               double rolloff);
+               WorkspaceBase* w);
 };
 
-struct BandStopBase : PoleFilterBase <AnalogLowPass>
+struct COMMON_LIB BandStopBase : PoleFilterBase <AnalogLowPass>
 {
     void setup(int order,
                double sampleRate,
                double centerFrequency,
                double widthFrequency,
-               double rippleDb,
-               double rolloff);
+               WorkspaceBase* w);
 };
 
 //------------------------------------------------------------------------------
@@ -175,21 +208,65 @@ struct BandStopBase : PoleFilterBase <AnalogLowPass>
 template <int MaxOrder>
 struct LowPass : PoleFilter <LowPassBase, MaxOrder>
 {
+    void setup(int order,
+               double sampleRate,
+               double cutoffFrequency)
+    {
+        Workspace <MaxOrder> w;
+        LowPassBase::setup(order,
+                           sampleRate,
+                           cutoffFrequency,
+                           &w);
+    }
 };
 
 template <int MaxOrder>
 struct HighPass : PoleFilter <HighPassBase, MaxOrder>
 {
+    void setup(int order,
+               double sampleRate,
+               double cutoffFrequency)
+    {
+        Workspace <MaxOrder> w;
+        HighPassBase::setup(order,
+                            sampleRate,
+                            cutoffFrequency,
+                            &w);
+    }
 };
 
 template <int MaxOrder>
 struct BandPass : PoleFilter <BandPassBase, MaxOrder, MaxOrder*2>
 {
+    void setup(int order,
+               double sampleRate,
+               double centerFrequency,
+               double widthFrequency)
+    {
+        Workspace <MaxOrder> w;
+        BandPassBase::setup(order,
+                            sampleRate,
+                            centerFrequency,
+                            widthFrequency,
+                            &w);
+    }
 };
 
 template <int MaxOrder>
 struct BandStop : PoleFilter <BandStopBase, MaxOrder, MaxOrder*2>
 {
+    void setup(int order,
+               double sampleRate,
+               double centerFrequency,
+               double widthFrequency)
+    {
+        Workspace <MaxOrder> w;
+        BandStopBase::setup(order,
+                            sampleRate,
+                            centerFrequency,
+                            widthFrequency,
+                            &w);
+    }
 };
 
 //------------------------------------------------------------------------------
@@ -205,27 +282,17 @@ struct TypeIBase : DesignBase
 {
     enum
     {
-        NumParams = 5
+        NumParams = 3
     };
 
     static int getNumParams()
     {
-        return 5;
+        return 3;
     }
 
     static const ParamInfo getParamInfo_2()
     {
         return ParamInfo::defaultCutoffFrequencyParam();
-    }
-
-    static const ParamInfo getParamInfo_3()
-    {
-        return ParamInfo::defaultRippleDbParam();
-    }
-
-    static const ParamInfo getParamInfo_4()
-    {
-        return ParamInfo::defaultRolloffParam();
     }
 };
 
@@ -234,7 +301,7 @@ struct TypeI : TypeIBase, FilterClass
 {
     void setParams(const Params& params)
     {
-        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4]);
+        FilterClass::setup(int(params[1]), params[0], params[2]);
     }
 };
 
@@ -242,12 +309,12 @@ struct TypeIIBase : DesignBase
 {
     enum
     {
-        NumParams = 6
+        NumParams = 4
     };
 
     static int getNumParams()
     {
-        return 6;
+        return 4;
     }
 
     static const ParamInfo getParamInfo_2()
@@ -259,16 +326,6 @@ struct TypeIIBase : DesignBase
     {
         return ParamInfo::defaultBandwidthHzParam();
     }
-
-    static const ParamInfo getParamInfo_4()
-    {
-        return ParamInfo::defaultRippleDbParam();
-    }
-
-    static const ParamInfo getParamInfo_5()
-    {
-        return ParamInfo::defaultRolloffParam();
-    }
 };
 
 template <class FilterClass>
@@ -276,7 +333,7 @@ struct TypeII : TypeIIBase, FilterClass
 {
     void setParams(const Params& params)
     {
-        FilterClass::setup(int(params[1]), params[0], params[2], params[3], params[4], params[5]);
+        FilterClass::setup(int(params[1]), params[0], params[2], params[3]);
     }
 };
 
@@ -290,7 +347,7 @@ struct LowPassDescription
     }
     static const char* getName()
     {
-        return "Elliptic Low Pass";
+        return "Legendre Low Pass";
     }
 };
 
@@ -302,7 +359,7 @@ struct HighPassDescription
     }
     static const char* getName()
     {
-        return "Elliptic High Pass";
+        return "Legendre High Pass";
     }
 };
 
@@ -314,7 +371,7 @@ struct BandPassDescription
     }
     static const char* getName()
     {
-        return "Elliptic Band Pass";
+        return "Legendre Band Pass";
     }
 };
 
@@ -326,7 +383,7 @@ struct BandStopDescription
     }
     static const char* getName()
     {
-        return "Elliptic Band Stop";
+        return "Legendre Band Stop";
     }
 };
 
@@ -346,6 +403,7 @@ struct OrderBase : TypeClass <FilterClass <MaxOrder> >
 
     }
 };
+
 //------------------------------------------------------------------------------
 
 //
@@ -353,25 +411,25 @@ struct OrderBase : TypeClass <FilterClass <MaxOrder> >
 //
 
 template <int MaxOrder>
-struct LowPass : OrderBase <MaxOrder, TypeI, Elliptic::LowPass>,
+struct LowPass : OrderBase <MaxOrder, TypeI, Legendre::LowPass>,
         LowPassDescription
 {
 };
 
 template <int MaxOrder>
-struct HighPass : OrderBase <MaxOrder, TypeI, Elliptic::HighPass>,
+struct HighPass : OrderBase <MaxOrder, TypeI, Legendre::HighPass>,
         HighPassDescription
 {
 };
 
 template <int MaxOrder>
-struct BandPass : OrderBase <MaxOrder, TypeII, Elliptic::BandPass>,
+struct BandPass : OrderBase <MaxOrder, TypeII, Legendre::BandPass>,
         BandPassDescription
 {
 };
 
 template <int MaxOrder>
-struct BandStop : OrderBase <MaxOrder, TypeII, Elliptic::BandStop>,
+struct BandStop : OrderBase <MaxOrder, TypeII, Legendre::BandStop>,
         BandStopDescription
 {
 };
