+++ Source/Plugins/CommonLibs/DspLib/MathSupplement.h
