+++ Source/Plugins/CommonLibs/DspLib/PoleFilter.cpp
