+++ Source/Plugins/CommonLibs/DspLib/RootFinder.cpp
