+++ Source/Plugins/CommonLibs/DspLib/RootFinder.h
