+++ Source/Plugins/CommonLibs/DspLib/SmoothedFilter.h
