+++ Source/Plugins/CommonLibs/DspLib/State.h
